+++ client/hooks/state/user/store.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { create } from "zustand"
 
 interface IUserState {
