+++ client/modules/main/components/carousels/contact-request.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import React from "react"
 import { useScopedI18n } from "@/utils/services/internationalization/client"
